pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jsmfsb"
version = "1.1.0"
authors = [
  { name="Darren Wilkinson", email="darrenjwilkinson@btinternet.com" },
]
description = "Python+JAX code relating to the textbook, Stochastic modelling for systems biology, third edition"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "sbmlsh",
  "numpy<2",
  "scipy",
  "jaxlib"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

[tool.hatch.build.targets.wheel]
packages = ["src/jsmfsb"]

[project.urls]
Homepage = "https://github.com/darrenjw/jax-smfsb"
Documentation = "https://jax-smfsb.readthedocs.io/"
Issues = "https://github.com/darrenjw/jax-smfsb/issues"

